Split Streamable HTTP client and server
